+++ packages/terra-core-docs/src/terra-dev-site/doc/hyperlink/example/ImageHyperlink.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import Hyperlink from 'terra-hyperlink';
 
-export default () => <Hyperlink href="https://www.cerner.com" variant="image">Image hyperlink</Hyperlink>;
+export default () => <Hyperlink text="Image hyperlink" href="https://www.cerner.com" variant="image" />;