+++ packages/terra-core-docs/src/terra-dev-site/doc/hyperlink/example/UnderlineHiddenHyperlink.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import Hyperlink from 'terra-hyperlink';
 
-export default () => <Hyperlink href="https://www.cerner.com" isUnderlineHidden>Underline hidden hyperlink</Hyperlink>;
+export default () => <Hyperlink text="Underline hidden hyperlink" href="https://www.cerner.com" isUnderlineHidden />;