+++ packages/terra-core-docs/src/terra-dev-site/test/hyperlink/DocumentHyperlink.test.jsx
@@ -3,6 +3,6 @@ import Hyperlink from 'terra-hyperlink';
 
 export default () => (
   <div role="main">
-    <Hyperlink id="link" href="https://www.cerner.com" variant="document">Document hyperlink</Hyperlink>
+    <Hyperlink text="Document hyperlink" id="link" href="https://www.cerner.com" variant="document" />
   </div>
 );