+++ packages/terra-core-docs/src/terra-dev-site/test/hyperlink/PrivateDisabledHyperlink.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Hyperlink from 'terra-hyperlink';
+
+export default () => (
+  <div role="main">
+    <Hyperlink text="Disabled hyperlink" href="https://www.cerner.com" isDisabled />
+    <br />
+    <Hyperlink text="Disabled icon hyperlink" href="https://www.cerner.com" isDisabled variant="document" />
+  </div>
+);